Don't process Sass partials
As written, the `permalink` compile option returning false will still
instruct Eleventy to process a template (in this case, a `_partial.scss`
file). This caused issues in SCSS files using `@import` and expecting
variables to be hoisted globally and available in all imported
stylesheets.

Assume an `application.scss`

```scss
@import "variables";
@import "elements"
```

…and a `_elements.scss` expecting that `$blue` is defined in
`_variables.scss`:

```scss
body {
  color: $blue;
}
```

Before this change, this setup would blow up the compiler as it tries to
process `_elements.scss` even though it's a partial.

So… moving the partial check code into the compile function. It'll
`return` and Eleventy will move on to the next file.